Added more refactoring, message callbacks and some stuff to the projectservice. Only need to bind the new data to the viewmodel now
