+++ ConsoleRunner/ConsoleRunner.csproj
@@ -59,6 +59,10 @@
       <Project>{ffe9979e-f1f6-439d-a60d-2cd1826a0202}</Project>
       <Name>PackageService</Name>
     </ProjectReference>
+    <ProjectReference Include="..\PackagesGeneral\PackagesGeneral.csproj">
+      <Project>{4075CA79-5685-404D-AAC0-4562F007F712}</Project>
+      <Name>PackagesGeneral</Name>
+    </ProjectReference>
     <ProjectReference Include="..\PackagesNpm\PackagesNpm.csproj">
       <Project>{7e7bd2cb-493f-45da-b23d-c8d79cb60f5b}</Project>
       <Name>PackagesNpm</Name>
