+++ Nuget/NugetPackage.cs
@@ -1,6 +1,6 @@
 ﻿using NuGet;
-using PackageMagic.PackageService.Interface;
-using PackageMagic.PackageService.Model;
+using PackageMagic.General.Interface;
+using PackageMagic.General.Type;
 using System.Collections.Generic;
 using System.Threading.Tasks;
