+++ PackageMagic.WPF/Design/DesignMagicProjectService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using PackageMagic.General.Type;
 using PackageMagic.PackageService.Interface;
 using PackageMagic.PackageService.Model;
