+++ PackageMagic.WPF/ViewModel/ViewModelLocator.cs
@@ -52,7 +52,7 @@ namespace PackageMagic.WPF.ViewModel
         
         public static void Cleanup()
         {
-            // TODO Clear the ViewModels
+            // INFO Clear the ViewModels here
         }
     }
 }
