+++ PackageManager/Service/MagicProjectService.cs
@@ -1,5 +1,9 @@
-﻿using PackageMagic.PackageService.Interface;
+﻿using PackageMagic.General.Interface;
+using PackageMagic.General.Type;
+using PackageMagic.Nuget;
+using PackageMagic.PackageService.Interface;
 using PackageMagic.PackageService.Model;
+using PackagesNpm;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -13,7 +17,7 @@ namespace PackageMagic.PackageService.Service
     {
         public MessageDelegate MessageCallback { get; set; }
 
-        public async Task<IEnumerable<IMagicProject>> GetProjectsAsync(string pathToSearch) => await Task.Run(() =>
+        public async Task<IEnumerable<IMagicProject>> GetProjectsAsync(string pathToSearch)
         {
             List<IMagicProject> listProjects = new List<IMagicProject>();
 
@@ -22,15 +26,28 @@ namespace PackageMagic.PackageService.Service
             {
                 MessageCallback?.Invoke($"Parsing {csProjFile}");
                 var project = new MagicProjectCs { Name = Path.GetFileName(csProjFile), Path = csProjFile };
-                //Create a nuget object for parsing nuget package references
-                //Create a npm object for parsing npm package references
-                //project.AddRange(await nuget.SearchPackages(project.Path));
-                //project.AddRange(await npm.SearchPackages(project.Path));
+                
+                //TODO! Bad naming convention forces the use of full namespace
+                //Consider using different namespace and class name for 'Nuget'
+                IMagicPackageSearch theSearcher = new Nuget.Nuget();
+                theSearcher.MessageCallback += MessageCallback;
+
+                //Use the nuget object for parsing nuget package references
+                project.Packages.AddRange(await theSearcher.SearchPackages(project.Path));
+                theSearcher.MessageCallback -= MessageCallback;
+
+                //TODO! Follow same naming conventions on namespaces and classnames in both Nuget and Npm library!
+                theSearcher = new Npm();
+                theSearcher.MessageCallback += MessageCallback;
+
+                //Use the npm object for parsing npm package references
+                project.Packages.AddRange(await theSearcher.SearchPackages(project.Path));
+                theSearcher.MessageCallback -= MessageCallback;
 
                 listProjects.Add(project);
             }
 
             return listProjects;
-        });
+        }
     }
 }
