+++ PackagesGeneral/PackagesGeneral.csproj
@@ -4,15 +4,14 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{FFE9979E-F1F6-439D-A60D-2CD1826A0202}</ProjectGuid>
+    <ProjectGuid>{4075CA79-5685-404D-AAC0-4562F007F712}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>PackageMagic.PackageService</RootNamespace>
-    <AssemblyName>PackageMagic.PackageService</AssemblyName>
+    <RootNamespace>PackageMagic.General</RootNamespace>
+    <AssemblyName>PackageMagic.General</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
-    <TargetFrameworkProfile />
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -42,26 +41,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Interface\IMagicProject.cs" />
-    <Compile Include="Interface\IMagicProjectService.cs" />
-    <Compile Include="Model\MagicProject.cs" />
-    <Compile Include="Service\MagicProjectService.cs" />
-    <Compile Include="Packages.cs" />
+    <Compile Include="Interface\IMagicPackage.cs" />
+    <Compile Include="Interface\IMagicPackageSearch.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Nuget\PackagesNuget.csproj">
-      <Project>{5B53AF09-68D2-4618-A13E-5F052945ADC5}</Project>
-      <Name>PackagesNuget</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\PackagesGeneral\PackagesGeneral.csproj">
-      <Project>{4075CA79-5685-404D-AAC0-4562F007F712}</Project>
-      <Name>PackagesGeneral</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\PackagesNpm\PackagesNpm.csproj">
-      <Project>{7E7BD2CB-493F-45DA-B23D-C8D79CB60F5B}</Project>
-      <Name>PackagesNpm</Name>
-    </ProjectReference>
+    <Compile Include="Type\Delegates.cs" />
+    <Compile Include="Type\MagickPackageType.cs" />
+    <Compile Include="Utils.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
