+++ PackagesGeneral/Type/Delegates.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PackageMagic.PackageService.Model
+namespace PackageMagic.General.Type
 {
     public delegate void MessageDelegate(string message);
 }
