+++ PackagesGeneral/Utils.cs
@@ -1,16 +1,11 @@
-﻿
-
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PackageMagic.PackageService
+namespace PackageMagic.General
 {
-
-
-
     public static class Utils
     {
         public static bool ExtendedLogging { get; private set; }
