+++ WinformsPackageMagic.sln
@@ -15,6 +15,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MVVM Application", "MVVM Ap
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PackageMagic.WPF", "PackageMagic.WPF\PackageMagic.WPF.csproj", "{C32AEF47-C7C0-4132-9805-02251490E7C8}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PackagesGeneral", "PackagesGeneral\PackagesGeneral.csproj", "{4075CA79-5685-404D-AAC0-4562F007F712}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,10 @@ Global
 		{C32AEF47-C7C0-4132-9805-02251490E7C8}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C32AEF47-C7C0-4132-9805-02251490E7C8}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C32AEF47-C7C0-4132-9805-02251490E7C8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4075CA79-5685-404D-AAC0-4562F007F712}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4075CA79-5685-404D-AAC0-4562F007F712}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4075CA79-5685-404D-AAC0-4562F007F712}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4075CA79-5685-404D-AAC0-4562F007F712}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
